Fix linter: Ineffectual assignment to prompt
